requirements.txt
pyrogram==2.0.106
pandas==2.1.4
python-dotenv==1.0.0
tgcrypto==1.2.5
aiohttp>=3.8.0 